TileLayer::parseCSV: use std::strtoul instead of std::stringstream
iostreams provide a nice interface to read numbers, but unfortunately they
incur a lot of extra work so they tend to be slow, so use std::strtoul directly
on the string.
I also added a reserve to the IDs vector, since the string ought to contain
width*height tiles, it also helps a bit.

Performance measurements (time to run ParseTest) on a 4096x4096 map with two
layers:

* clang with libc++:
  original: 6.209s, strtoull: 0.773s, strtoull+reserve: 0.723s
* clang with libstdc++:
  original: 2.153s, strtoull: 0.800s, strtoull+reserve: 0.736s
* gcc:
  original: 2.115s, strtoull: 0.805s, strtoull+reserve: 0.752s