Add meson file for ParseTest
Also fix the relative path for which the
test wouldn't find the required assets.
Add a build-time option to disable the
unit test's request for user input at the
end, when it does it breaks the test if
run with `ninja test`. Leaving enabled by
default as this is what the project
currently does.